Implementa cadastro de anotacoes por tela integrada
