+++ app/screens/featurePaciente/registrosPaciente/components/RelatoModal.tsx
@@ -1,7 +1,7 @@
-// components/RelatoModal.tsx
 import { AnotacaoPacienteModel } from '@/constants/models/AnotacaoPacienteModel';
+import styles from '@/styles/RelatoModalStyle';
 import React from 'react';
-import { Modal, View, Text, StyleSheet, TouchableOpacity } from 'react-native';
+import { Modal, View, Text, TouchableOpacity } from 'react-native';
 
 interface RelatoModalProps {
   visivel: boolean;
@@ -30,50 +30,6 @@ const RelatoModal: React.FC<RelatoModalProps> = ({ visivel, fechar, registro }) 
   );
 };
 
-const styles = StyleSheet.create({
-  overlay: {
-    flex: 1,
-    backgroundColor: '#00000099',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  container: {
-    width: '85%',
-    backgroundColor: '#fff',
-    padding: 20,
-    borderRadius: 10,
-    elevation: 5,
-  },
-  closeButton: {
-    alignSelf: 'flex-end',
-  },
-  closeText: {
-    color: '#007AFF',
-    fontWeight: 'bold',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginVertical: 10,
-  },
-  status: {
-    color: 'gray',
-    marginBottom: 10,
-  },
-  texto: {
-    fontSize: 16,
-    marginBottom: 15,
-  },
-  categoriesContainer: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-  },
-  category: {
-    backgroundColor: '#eee',
-    padding: 5,
-    borderRadius: 5,
-    margin: 3,
-  },
-});
+
 
 export default RelatoModal;
