+++ constants/models/AnotacaoPacienteModel.ts
@@ -1,43 +1,26 @@
 export class AnotacaoPacienteModel {
-    public idAnotacao: number;
-    public descricao: string;
-    public emocaoEstimada: string | null;
-    public dhRegistro: Date;
-    public _fk_idPaciente: number;
-    public titulo: string;
-    public isVisualizada: boolean
-
-    constructor(
-        idAnotacao: number,
-        descricao: string,
-        dhRegistro: Date,
-        fk_idPaciente: number,
-        emocaoEstimada: string | null = null,
-        titulo:  string,
-        isVisualizada: boolean,
-    ) {
-        this.idAnotacao = idAnotacao;
-        this.descricao = descricao;
-        this.dhRegistro = dhRegistro;
-        this._fk_idPaciente = fk_idPaciente;
-        this.emocaoEstimada = emocaoEstimada;
-        this.titulo = titulo;
-        this.isVisualizada = isVisualizada
+    public idAnotacao: number = 0;
+    public descricao: string = '';
+    public emocaoEstimada: string | null = null;
+    public dhRegistro: Date = new Date();
+    public _fk_idPaciente: number = 0;
+    public titulo: string = '';
+    public isVisualizada: boolean = false;
+  
+    constructor() {
+      // Construtor vazio: todas as propriedades são inicializadas acima
     }
-    
-
-      
+  
     public static copy(anotacao: AnotacaoPacienteModel): AnotacaoPacienteModel {
-        return new AnotacaoPacienteModel(
-        anotacao.idAnotacao,
-        anotacao.descricao,
-        anotacao.dhRegistro,
-        anotacao._fk_idPaciente,
-        anotacao.emocaoEstimada,
-        anotacao.titulo,
-        anotacao.isVisualizada
-        );
+      const copia = new AnotacaoPacienteModel();
+      copia.idAnotacao = anotacao.idAnotacao;
+      copia.descricao = anotacao.descricao;
+      copia.emocaoEstimada = anotacao.emocaoEstimada;
+      copia.dhRegistro = anotacao.dhRegistro;
+      copia._fk_idPaciente = anotacao._fk_idPaciente;
+      copia.titulo = anotacao.titulo;
+      copia.isVisualizada = anotacao.isVisualizada;
+      return copia;
     }
-
-
-}
+  }
+
