+++ constants/models/UsuarioLogadoContextProps.ts
@@ -1,0 +1,7 @@
+import { UsuarioLogado } from "./UsuarioLogado";
+
+export default interface UsuarioLogadoContextProps {
+    usuarioLogado: UsuarioLogado;
+    setUsuarioLogado: (usuario: UsuarioLogado) => void;
+    getUsuarioLogado: () => UsuarioLogado;
+}
