+++ constants/models/UsuarioLogadoProviderProps.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export default interface UsuarioLogadoProviderProps {
+  children: ReactNode;
+}
