+++ util/DateUtil.ts
@@ -42,4 +42,42 @@ export class DateUtil {
               date.getMonth() + 1
             ).padStart(2, '0')}/${date.getFullYear()}`;
         };
+
+        public static formatarData = (texto: string) => {
+            let cleaned = texto.replace(/\D/g, '');
+            if (cleaned.length > 8) {
+              cleaned = cleaned.slice(0, 8);
+            }
+            if (cleaned.length >= 5) {
+              return `${cleaned.slice(0, 2)}/${cleaned.slice(2, 4)}/${cleaned.slice(4)}`;
+            } else if (cleaned.length >= 3) {
+              return `${cleaned.slice(0, 2)}/${cleaned.slice(2)}`;
+            } else {
+              return cleaned;
+            }
+          };
+          
+        public static validarData = (dataTexto: string) => {
+            const [dia, mes, ano] = dataTexto.split('/').map(Number);
+            const data = new Date(ano, mes - 1, dia);
+            if (data.getFullYear() !== ano || data.getMonth() !== mes - 1 || data.getDate() !== dia) {
+              return { valido: false, mensagem: 'Data inválida' };
+            }
+            const hoje = new Date();
+            if (data > hoje) {
+              return { valido: false, mensagem: 'Data no futuro' };
+            }
+            return { valido: true };
+          };
+          
+        public static validarIntervaloDatas = (dataInicioTexto: string, dataFimTexto: string) => {
+            const [diaInicio, mesInicio, anoInicio] = dataInicioTexto.split('/').map(Number);
+            const [diaFim, mesFim, anoFim] = dataFimTexto.split('/').map(Number);
+            const dataInicio = new Date(anoInicio, mesInicio - 1, diaInicio);
+            const dataFim = new Date(anoFim, mesFim - 1, diaFim);
+            if (dataFim < dataInicio) {
+              return { valido: false, mensagem: 'Data final é menor que a data inicial' };
+            }
+            return { valido: true , mensagem : 'Valido'};
+        };
 }
